requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose==3.3.0
PyJWT==2.10.1
requests==2.32.3
beautifulsoup4==4.13.3
lxml==5.3.1
jinja2==3.1.5
python-multipart==0.0.20
schedule==1.2.2
chardet==5.2.0
